Changed type of partID from int to long

+++ app/src/main/java/edu/chalmers/sikkr/backend/mms/MMSInbox.java
@@ -46,7 +46,7 @@ public class MMSInbox {
         Cursor cursor = context.getContentResolver().query(CONTENT_URI, null, null, null, null);
         while (cursor.moveToNext()) {
             String sender = cursor.getString(cursor.getColumnIndexOrThrow(ADDRESS));
-            int partID = cursor.getInt(cursor.getColumnIndexOrThrow(_ID));
+            long partID = cursor.getLong(cursor.getColumnIndexOrThrow(_ID));
 
             Cursor curPart = context.getContentResolver().query(Uri.parse ("content://mms/" + partID + "/part"), null, null, null, null);
             curPart.moveToFirst();
